+++ src/main/java/janggi/board/BoardFactory.java
@@ -1,4 +1,4 @@
-package janggi;
+package janggi.board;
 
 import janggi.piece.Byeong;
 import janggi.piece.Cannon;
@@ -9,6 +9,8 @@ import janggi.piece.Horse;
 import janggi.piece.Jol;
 import janggi.piece.King;
 import janggi.piece.Piece;
+import janggi.piece.Team;
+import janggi.position.Position;
 import java.util.HashMap;
 import java.util.Map;
 import java.util.Map.Entry;
