+++ src/main/java/janggi/piece/Team.java
@@ -1,4 +1,4 @@
-package janggi;
+package janggi.piece;
 
 public enum Team {
