+++ src/main/java/janggi/position/Path.java
@@ -1,4 +1,4 @@
-package janggi;
+package janggi.position;
 
 import java.util.List;
 import java.util.Objects;
