+++ src/test/java/janggi/PositionTest.java
@@ -1,5 +1,6 @@
 package janggi;
 
+import janggi.position.Position;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
