+++ src/test/java/janggi/piece/ByeongTest.java
@@ -1,7 +1,7 @@
 package janggi.piece;
 
-import janggi.Path;
-import janggi.Position;
+import janggi.position.Path;
+import janggi.position.Position;
 import java.util.List;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
